+++ retz-common/src/test/java/io/github/retz/protocol/ProtocolTest.java
@@ -51,21 +51,21 @@ public class ProtocolTest {
     @Test
     public void decodeObjects() throws IOException {
         {
-            String jobString = "{ \"appid\":\"myid\", \"cmd\":\"yaess/bin/yaess-batch.sh ....\", \"scheduled\":\"2016-05-15:20:20:20Z\", \"started\":\"2016-05-15:20:20:20Z\", \"finished\":\"2016-05-15:20:20:20Z\", \"result\":0, \"id\":1, \"url\":\"https://....\", \"state\":\"STARTED\"}";
+            String jobString = "{ \"appid\":\"myid\", \"cmd\":\"yaess/bin/yaess-batch.sh ....\", \"scheduled\":\"2016-05-15:20:20:20Z\", \"started\":\"2016-05-15:20:20:20Z\", \"finished\":\"2016-05-15:20:20:20Z\", \"result\":0, \"cpu\":22,\"memMB\":234,\"id\":1, \"url\":\"https://....\", \"state\":\"STARTED\"}";
             Job job = mapper.readValue(jobString, Job.class);
             assertThat(job.cmd(), is("yaess/bin/yaess-batch.sh ...."));
             assertThat(job.result(), is(0));
             assertThat(job.id(), is(1));
         }
         {
-            String jobString = "{ \"appid\" : \"myid\", \"cmd\":\"yaess/bin/yaess-batch.sh ....\", \"scheduled\":\"2016-05-15:20:20:20Z\", \"id\":2, \"state\":\"STARTED\"}";
+            String jobString = "{ \"appid\" : \"myid\", \"cmd\":\"yaess/bin/yaess-batch.sh ....\", \"scheduled\":\"2016-05-15:20:20:20Z\", \"id\":2, \"state\":\"STARTED\", \"cpu\":1,\"memMB\":128}";
             Job job = mapper.readValue(jobString, Job.class);
             assertThat(job.cmd(), is("yaess/bin/yaess-batch.sh ...."));
             assertNull(job.started());
             assertThat(job.result(), is(0));
         }
         {
-            String appString = "{\"appid\":\"my-one-of-42-apps\", \"files\":[\"http://example.com/app.tar.gz\"], \"container\":{\"type\":\"mesos\"}, \"owner\":\"deadbeef\" }";
+            String appString = "{\"appid\":\"my-one-of-42-apps\", \"files\":[\"http://example.com/app.tar.gz\"], \"container\":{\"type\":\"mesos\"}, \"owner\":\"deadbeef\"}";
             Application app = mapper.readValue(appString, Application.class);
             assertThat(app.getAppid(), is("my-one-of-42-apps"));
             assertNotNull(app.getFiles());
@@ -79,7 +79,7 @@ public class ProtocolTest {
             assertFalse(app.getDiskMB().isPresent());
         }
         {
-            String metajobString = "{\"job\":{\"cmd\":\"touch /tmp/junit3807420247585460493.tmp\",\"scheduled\":null,\"started\":null,\"finished\":null,\"result\":-1,\"id\":0,\"url\":null,\"appid\":\"appname\",\"name\":null,\"cpu\":{\"min\":1,\"max\":2147483647},\"memMB\":{\"min\":128,\"max\":2147483647},\"props\":null, \"state\":\"STARTED\"},\"app\":{\"appid\":\"appname\",\"persistentFiles\":[],\"files\":[],\"diskMB\":null, \"container\":{\"type\":\"mesos\"}, \"owner\":\"deadbeef\"}}";
+            String metajobString = "{\"job\":{\"cmd\":\"touch /tmp/junit3807420247585460493.tmp\",\"scheduled\":null,\"started\":null,\"finished\":null,\"result\":-1,\"id\":0,\"url\":null,\"appid\":\"appname\",\"name\":null,\"props\":null, \"state\":\"STARTED\", \"cpu\":1,\"memMB\":128},\"app\":{\"appid\":\"appname\",\"persistentFiles\":[],\"files\":[],\"diskMB\":null, \"container\":{\"type\":\"mesos\"}, \"owner\":\"deadbeef\"}}";
             MetaJob metaJob = mapper.readValue(metajobString, MetaJob.class);
             assertNotNull(metaJob.getApp());
         }
@@ -94,7 +94,7 @@ public class ProtocolTest {
             assertThat(req, instanceOf(ListJobRequest.class));
         }
         {
-            String json = "{\"command\":\"schedule\",\"job\":{\"cmd\":\"ls -l\", \"id\":0,\"url\":null,\"appid\":\"appname\", \"state\":\"QUEUED\"}}";
+            String json = "{\"command\":\"schedule\",\"job\":{\"cmd\":\"ls -l\", \"id\":0,\"url\":null,\"cpu\":23,\"memMB\":324,\"appid\":\"appname\", \"state\":\"QUEUED\"}}";
             Request req = mapper.readValue(json, Request.class);
             assertThat(req, instanceOf(ScheduleRequest.class));
             ScheduleRequest sreq = (ScheduleRequest) req;
@@ -102,7 +102,7 @@ public class ProtocolTest {
             assertThat(sreq.job().cmd(), is("ls -l"));
         }
         {
-            String json = "{\"command\":\"schedule\",\"job\":{\"cmd\":\"Mmmmmmmmmy commmmmand1!!!!!\",\"scheduled\":null,\"appid\":\"aaa\",\"finished\":null,\"result\":0,\"id\":0,\"url\":null,  \"state\":\"QUEUED\"}}";
+            String json = "{\"command\":\"schedule\",\"job\":{\"cmd\":\"Mmmmmmmmmy commmmmand1!!!!!\",\"scheduled\":null,\"appid\":\"aaa\",\"finished\":null,\"result\":0,\"id\":0,\"url\":null, \"cpu\":12,\"memMB\":23445, \"state\":\"QUEUED\"}}";
             Request req = mapper.readValue(json, Request.class);
             assertThat(req, instanceOf(ScheduleRequest.class));
             ScheduleRequest sreq = (ScheduleRequest) req;
@@ -141,7 +141,7 @@ public class ProtocolTest {
             assertThat(res, instanceOf(KillResponse.class));
         }
         {
-            String jobJson = "{\"appid\" : \"moo\", \"cmd\":\"ls -l\", \"id\":23, \"state\":\"STARTED\"}";
+            String jobJson = "{\"appid\" : \"moo\", \"cmd\":\"ls -l\", \"cpu\":123,\"memMB\":234,\"id\":23, \"state\":\"STARTED\"}";
             String json = "{\"command\":\"schedule\", \"status\":\"ok\", \"job\":" + jobJson + "}";
             Response res = mapper.readValue(json, Response.class);
             assertThat(res, instanceOf(ScheduleResponse.class));
@@ -150,7 +150,7 @@ public class ProtocolTest {
             assertThat(sres.job.id(), is(23));
         }
         {
-            String json = "{\"command\":\"get-job\",\"job\":{\"appid\":\"foobar\", \"cmd\":\"Mmmmmmmmmy commmmmand1!!!!!\",\"scheduled\":null,\"started\":null,\"finished\":null,\"result\":0,\"id\":0,\"url\":null, \"state\":\"STARTED\"}}";
+            String json = "{\"command\":\"get-job\",\"job\":{\"appid\":\"foobar\", \"cmd\":\"Mmmmmmmmmy commmmmand1!!!!!\",\"scheduled\":null,\"started\":null,\"finished\":null,\"result\":0,\"id\":0,\"url\":null, \"cpu\":1,\"memMB\":128, \"state\":\"STARTED\"}}";
 
             Response res = mapper.readValue(json, Response.class);
             assertThat(res, instanceOf(GetJobResponse.class));
@@ -159,7 +159,7 @@ public class ProtocolTest {
             assertThat(getJobResponse.job().get().cmd(), is("Mmmmmmmmmy commmmmand1!!!!!"));
         }
         {
-            String json = "{\"command\":\"watch\", \"status\":\"ok\", \"event\":\"finished\", \"job\":{\"cmd\":\"ls\", \"id\":23, \"appid\":\"myapp\", \"state\":\"STARTED\"}}";
+            String json = "{\"command\":\"watch\", \"status\":\"ok\", \"event\":\"finished\", \"job\":{\"cmd\":\"ls\", \"id\":23, \"appid\":\"myapp\", \"state\":\"STARTED\", \"cpu\":1,\"memMB\":128}}";
             Response res = mapper.readValue(json, Response.class);
             assertThat(res, instanceOf(WatchResponse.class));
             WatchResponse wres = (WatchResponse) res;
@@ -187,7 +187,7 @@ public class ProtocolTest {
         }
 
         {
-            Job job = new Job("foobar-app", "ls -l", null, new Range(1, 0), new Range(128, 0));
+            Job job = new Job("foobar-app", "ls -l", null, 1, 128);
             ScheduleRequest scheduleRequest = new ScheduleRequest(job, false);
             String json = mapper.writeValueAsString(scheduleRequest);
             Request req = mapper.readValue(json, Request.class);